revert(antd): add useSider and share logic between siders
This reverts commit 4c7547fa1b91c38fafb4fb26d2ab5be9e96686b8.
