+++ packages/antd/src/hooks/index.ts
@@ -5,4 +5,3 @@ export * from "./import";
 export * from "./list";
 export * from "./useFileUploadState";
 export * from "./modal";
-export * from "./layout";
