--- packages/antd/src/hooks/layout/index.ts
@@ -1,1 +1,0 @@
-export * from "./useSider";
